Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
Flask-JWT-Extended
Flask-CORS
Flask-Migrate
bcrypt  # For password hashing
python-dotenv  # If you're using a .env file for configuration
